+++ ChangeLog.md
@@ -19,6 +19,7 @@
 - Fixed vecWeiPixRadT extraction in ShSlopec
 - Fixed output\_names in PhaseScreenCube
 - Fixed start\_time bug in WindowedIntegration
+- Optimization of the compute\_ifs\_covmat function
 - ...
 
 ## [1.0.3] - 2026-05-18